requirements.txt
google-search-results
requests
beautifulsoup4
langchain
faiss-cpu
sentence-transformers
langchain-community
python-dotenv
sarvamai
torch
pytorch-lightning
gradio
fastapi
uvicorn
# NeMo is installed via git clone in Colab